Move deprecated blocks to correct directory
