+++ private/src/scripts/blocks.js
@@ -15,19 +15,10 @@ import './editor/block-styles';
 import './editor/plugins/appearance-options/index.jsx';
 import './editor/plugins/pop-in/index.jsx';
 
-import './editor/blocks-deprecated/banner/index.jsx';
-import './editor/blocks-deprecated/header/index.jsx';
-
 import './editor/blocks/action/index.jsx';
-import './editor/blocks/background-media/index.jsx';
 import './editor/blocks/blockquote/index.jsx';
-import './editor/blocks/button/index.jsx';
-import './editor/blocks/call-to-action/index.jsx';
-import './editor/blocks/collapsable/index.jsx';
-import './editor/blocks/columns/index.jsx';
 import './editor/blocks/core-mods';
 import './editor/blocks/countdown-timer/index.jsx';
-import './editor/blocks/custom-card/index.jsx';
 import './editor/blocks/download/index.jsx';
 import './editor/blocks/embed-flourish/index.jsx';
 import './editor/blocks/embed-infogram/index.jsx';
@@ -37,10 +28,8 @@ import './editor/blocks/hero/index.jsx';
 import './editor/blocks/hero/replaceHeaders';
 import './editor/blocks/iframe-button/index.jsx';
 import './editor/blocks/iframe/index.jsx';
-import './editor/blocks/image/index.jsx';
 import './editor/blocks/key-facts/index.jsx';
 import './editor/blocks/link-group/index.jsx';
-import './editor/blocks/links-with-icons/index.jsx';
 import './editor/blocks/menu/index.jsx';
 import './editor/blocks/petition-list/index.jsx';
 import './editor/blocks/post-list/index.jsx';
@@ -55,3 +44,15 @@ import './editor/blocks/tweet/index.jsx';
 
 import './editor/fse-blocks/pop-in/index.jsx';
 import './editor/fse-blocks/site-header/index.jsx';
+
+// To remove in v2.0.0
+import './editor/blocks-deprecated/background-media/index.jsx';
+import './editor/blocks-deprecated/banner/index.jsx';
+import './editor/blocks-deprecated/button/index.jsx';
+import './editor/blocks-deprecated/call-to-action/index.jsx';
+import './editor/blocks-deprecated/collapsable/index.jsx';
+import './editor/blocks-deprecated/columns/index.jsx';
+import './editor/blocks-deprecated/custom-card/index.jsx';
+import './editor/blocks-deprecated/header/index.jsx';
+import './editor/blocks-deprecated/image/index.jsx';
+import './editor/blocks-deprecated/links-with-icons/index.jsx';
