+++ private/src/scripts/editor/blocks-deprecated/background-media/DisplayComponent.jsx
