+++ private/src/scripts/editor/blocks-deprecated/background-media/components/deprecated.jsx
