+++ private/src/scripts/editor/blocks-deprecated/background-media/deprecated.jsx
