+++ private/src/scripts/editor/blocks-deprecated/background-media/index.jsx
