+++ private/src/scripts/editor/blocks-deprecated/collapsable/DisplayComponent.jsx
