+++ private/src/scripts/editor/blocks-deprecated/collapsable/index.jsx
