+++ private/src/scripts/editor/blocks-deprecated/columns/DisplayComponent.jsx
