+++ private/src/scripts/editor/blocks-deprecated/image/DisplayComponent.jsx
