+++ private/src/scripts/editor/blocks-deprecated/image/deprecated.jsx
