+++ private/src/scripts/editor/blocks-deprecated/image/utils.js
