+++ private/src/scripts/editor/blocks-deprecated/links-with-icons/DisplayComponent.jsx
