+++ private/src/scripts/editor/blocks-deprecated/links-with-icons/InnerDisplayComponent.jsx
